requirements.txt
langchain
chromadb
sentence-transformers
requests
python-dotenv
google-generativeai